refactor: display component to widgets chore: upgrade tsc
