+++ package.json
@@ -33,6 +33,6 @@
     "lerna": "^3.16.4",
     "lint-staged": "^10.2.11",
     "prettier": "^1.19.1",
-    "typescript": "^3.9.3"
+    "typescript": "^3.9.7"
   }
 }
