+++ packages/cli/package.json
@@ -12,9 +12,9 @@
     "pub": "npm publish --access public"
   },
   "dependencies": {
-    "@babel/preset-typescript": "^7.9.0",
     "@babel/core": "^7.9.0",
     "@babel/plugin-transform-async-to-generator": "^7.8.3",
+    "@babel/preset-typescript": "^7.9.0",
     "babel-loader": "^8.1.0",
     "chalk": "^3.0.0",
     "commander": "^5.0.0",
@@ -29,7 +29,7 @@
     "raw-loader": "^4.0.0",
     "sass-loader": "^8.0.2",
     "string.prototype.padstart": "^3.1.0",
-    "typescript": "^3.8.3",
+    "typescript": "^4.0.2",
     "urllib": "^2.34.2",
     "webpack": "^4.42.1",
     "webpack-dev-server": "^3.10.3",
