+++ packages/editor/package.json
@@ -28,11 +28,10 @@
     "react-scripts": "3.4.1",
     "react-sortable-hoc": "^1.11.0",
     "react-use": "^13.27.1",
+    "react-rnd": "^10.1.10",
     "styled-components": "^5.1.1",
     "throttle-debounce": "^2.1.0",
-    "tinycolor2": "^1.4.1",
-    "typescript": "^3.8.3",
-    "webpack": "4.42.0"
+    "tinycolor2": "^1.4.1"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.10",
@@ -47,7 +46,8 @@
     "customize-cra": "^0.9.1",
     "node-sass": "^4.13.1",
     "react-app-rewired": "^2.1.6",
-    "react-rnd": "^10.1.10"
+    "typescript": "^4.0.2",
+    "webpack": "4.42.0"
   },
   "browserslist": {
     "production": [
