+++ packages/editor/src/components/MaterialsView/MaterialsLibrary/ComponentsLibrary/MaterialsComponentPreview.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { MaterialsComponentMeta } from 'types';
-import { SVGRender } from 'components/SVGRender';
+import { SVGRender } from 'widgets/SVGRender';
 
 interface Props {
   item: MaterialsComponentMeta;
