+++ packages/editor/src/widgets/EditableText/index.scss
