+++ packages/editor/src/widgets/Empty/index.scss
