+++ packages/editor/src/widgets/ErrorBoundary/index.tsx
