+++ packages/editor/src/widgets/SVGRender/index.tsx
