+++ packages/editor/src/widgets/Simulator/index.scss
