+++ packages/materials-universal/package.json
@@ -1,43 +1,43 @@
 {
-    "name": "@vize/materials-universal",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "dev:local": "/Users/huqingyang/Desktop/Proj/vize/packages/cli/bin/vize.js dev ./"
-    },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "antd-mobile": "^2.3.1",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1"
-    },
-    "devDependencies": {
-        "antd-mobile": "^2.3.1",
-        "@babel/plugin-proposal-class-properties": "^7.5.5",
-        "@babel/plugin-proposal-decorators": "^7.4.4",
-        "@babel/plugin-proposal-export-default-from": "^7.5.2",
-        "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
-        "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-        "@babel/plugin-transform-async-to-generator": "^7.5.0",
-        "@babel/plugin-transform-runtime": "^7.5.5",
-        "@babel/polyfill": "^7.6.0",
-        "@babel/preset-env": "^7.4.4",
-        "@babel/preset-react": "^7.0.0",
-        "@babel/preset-typescript": "^7.9.0",
-        "@types/node": "^13.9.8",
-        "@types/react": "^16.9.29",
-        "@types/react-dom": "^16.9.5",
-        "babel-loader": "^8.1.0",
-        "base64-inline-loader": "^1.1.1",
-        "dart-sass": "^1.25.0",
-        "raw-loader": "^4.0.0",
-        "css-loader": "^3.4.2",
-        "sass-loader": "^8.0.2",
-        "typescript": "^3.8.3",
-        "watchpack": "^1.6.1",
-        "webpack": "^4.42.1",
-        "webpack-cli": "^3.3.11"
-    }
+  "name": "@vize/materials-universal",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev:local": "/Users/huqingyang/Desktop/Proj/vize/packages/cli/bin/vize.js dev ./"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "antd-mobile": "^2.3.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
+  },
+  "devDependencies": {
+    "antd-mobile": "^2.3.1",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/plugin-proposal-export-default-from": "^7.5.2",
+    "@babel/plugin-proposal-export-namespace-from": "^7.5.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
+    "@babel/plugin-transform-async-to-generator": "^7.5.0",
+    "@babel/plugin-transform-runtime": "^7.5.5",
+    "@babel/polyfill": "^7.6.0",
+    "@babel/preset-env": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.9.0",
+    "@types/node": "^13.9.8",
+    "@types/react": "^16.9.29",
+    "@types/react-dom": "^16.9.5",
+    "babel-loader": "^8.1.0",
+    "base64-inline-loader": "^1.1.1",
+    "dart-sass": "^1.25.0",
+    "raw-loader": "^4.0.0",
+    "css-loader": "^3.4.2",
+    "sass-loader": "^8.0.2",
+    "typescript": "^4.0.2",
+    "watchpack": "^1.6.1",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11"
+  }
 }
